requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
django-uuid-upload-path==1.0.0
django==1.9.4
pillow==3.1.1
